requirements.txt
anthropic==0.79.0
python-dotenv==1.0.0
pydantic==2.12.5

# RAG System Dependencies
# Note: Using chromadb 0.4.24 for Python 3.13 compatibility
# Newer versions require onnxruntime which doesn't support Python 3.13 yet
chromadb==0.4.24
openai>=1.10.0
tiktoken>=0.5.2